refactor: remove index from types dir
- Allows for more delcaritive and easier to understand type imports
