--- src/types/index.ts
@@ -1,5 +1,0 @@
-export * from './common'
-export * from './functions'
-export * from './guards'
-export * from './responsive'
-export * from './theme'
